requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==37.0.4
Deprecated==1.2.13
Django==4.1.1
gcloud==0.18.3
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna==3.3
jwcrypto==1.3.1
oauth2client==4.1.3
protobuf==4.21.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
pyparsing==3.0.9
Pyrebase4==4.5.0
python-jwt==3.3.4
requests==2.28.1
requests-toolbelt==0.9.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
wrapt==1.14.1
